mas cambios
otros

+++ database/migrations/2017_11_15_020610_create_services_table.php
@@ -16,7 +16,7 @@ class CreateServicesTable extends Migration
         // Los Servicios que tiene la Clínica a los que van los Equipo
         Schema::create('services', function (Blueprint $table) {
             $table->increments('id');
-            $table->numeric('ceco');
+            $table->integer('ceco');
             $table->string('name');
             $table->timestamps();
         });
