+++ database/migrations/2017_11_15_020733_create_replacements_table.php
@@ -17,7 +17,7 @@ class CreateReplacementsTable extends Migration
         Schema::create('replacements', function (Blueprint $table) {
             $table->increments('id');
             $table->string('numero_parte');
-            $table->numeric('cantidad');
+            $table->integer('cantidad');
             $table->string('descripcion');
             $table->timestamps();
         });
